server/requirements.txt
absl-py==2.1.0
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cryptography
cycler==0.12.1
databases==0.9.0
fastapi==0.115.4
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-texttospeech==2.20.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.0
grpcio-status==1.62.3
h11==0.14.0
huggingface-hub==0.26.1
idna==3.10
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mediapipe==0.10.14
ml_dtypes==0.5.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
passlib==1.7.4
passlib[bcrypt]
pillow==11.0.0
proto-plus==1.25.0
protobuf==4.25.5
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydantic[email]
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv
python-jose
python-multipart==0.0.16
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.36
starlette==0.41.2
sympy==1.13.1
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.46.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
